refactor(*): move CancelableEventArgs to core; remove unnecessary tick calls
