+++ projects/igniteui-angular/src/lib/services/overlay/utilities.ts
@@ -5,6 +5,7 @@ import { IScrollStrategy, NoOpScrollStrategy } from './scroll';
 import { AnimationMetadata, AnimationReferenceMetadata, AnimationPlayer } from '@angular/animations';
 import { ComponentRef, ElementRef } from '@angular/core';
 import { IgxOverlayOutletDirective } from '../../directives/toggle/toggle.directive';
+import { CancelableEventArgs } from '../../core/utils';
 
 export enum HorizontalAlignment {
     Left = -1,
@@ -47,11 +48,7 @@ export interface OverlayEventArgs {
     componentRef?: ComponentRef<{}>;
 }
 
-export interface OverlayCancelableEventArgs extends OverlayEventArgs {
-    /**
-     * Provides the ability to cancel the event.
-     */
-    cancel: boolean;
+export interface OverlayCancelableEventArgs extends OverlayEventArgs, CancelableEventArgs {
 }
 
 export interface OverlayAnimationEventArgs {
